Move tests/core/usage/is_regular.h -> tests/include/sqlpp23/tests/core/
