+++ tests/include/sqlpp23/tests/core/is_regular.h
